+++ proto/cluster.go
@@ -1,6 +1,0 @@
-package proto
-
-type ClusterNode struct {
-	ID        string
-	Advertise string
-}