+++ proto/status.go
@@ -1,11 +1,22 @@
 package proto
 
-type NodeResponse struct {
-	Id        string
+type NodeInfo struct {
+	ID        string
 	Advertise string
-	Drain     string
 	Version   string
 	API       string
 }
 
-type NodesResponse []NodeResponse
+type NodesInfo []NodeInfo
+
+func (c NodesInfo) Len() int {
+	return len(c)
+}
+
+func (c NodesInfo) Less(i, j int) bool {
+	return c[i].ID < c[j].ID
+}
+
+func (c NodesInfo) Swap(i, j int) {
+	c[i], c[j] = c[j], c[i]
+}